Modify BotDebugger to return output instead of dumping it to stdout
This way we can use the BotDebugger methods and its return values
internall, resp. it can be used by another tool that wants to somehow
present the output differently (like an API).
One downside is, that the output is being `buffered`, i.e. *all* the
messages are only returned when the BotDebugger `run` method returns.

This also updates the `intelmqctl` script to use the new BotDebugger
interface.