Deprecate the XMPP bots
Both the XMPP output bot as well as the XMPP collector bot require
the `sleekxmpp` library which is deprecated. It would be possible to
modify the bots to use the successor `slixmpp` library, but a short
survey
https://lists.cert.at/pipermail/intelmq-users/2020-October/000177.html
did not bring up anyone using the bots, so they are being marked as
deprecated for now and will be removed in version 3.0.