Add a python script to post automatic codestyle reviews on PRs
This commit introduces two python modules in the
~/.github/workflows/python directory. The github.py module provides
basic functionality to interact with a pull request (posting a review,
listing existing reviews and updating an existing review).
The pycodestyle_comment.py module uses the github.py module to post a
review containing the pycodestyle output. On a new push it deletes any
earlier reviews it posted before.
The pycodestyle.yml action is used to run the pycodestyle_comment.py
script on every PR.