Adapt doc buildsystem to only produce restructured text
The script `docs/autogen.py` still generated the harmonization field
overview as markdown file. To get rid of the recommonmark dependency,
the script now generates restructured text, like all the other
documentaiton source files.
This commit modifies the autogen.py file as described, it changes the
extension of the autogenerated file in .gitignore; it removes all traces
of recommonmark configs in the sphinx configuration files docs/conf.py.
It updates the filename of the autogenerated file in the documentation
index.rst file and it drops the now unused dependencies from the
requirements.txt file.
The commit also drops the doc/dev/IntelMQ-3.0-Architecture.md file from
the documentation- this is an internal WIP file and does not need to be
part of the user facing docs.
Some of the field descriptions in the harmonization.conf file contained
markdown syntax, they were also updated.