Add an IntelMQ Shadowserver API collector and JSON parser
In addition to sending out emails, Shadowserver now provides a report
API. The Shadowserver API collector connects to the API and downloads
the reports from the last couple of days, checks which ones it has not
processed yet and passes those on. There is a now also a Shadowserver
JSON parser, that is able to parse the API reports.

Closes #1683