MAINT: Add a Github workflow that builds Debian packages
The workflow builds debian packages on Debian buster and bullseye. The
action runs on push & pull request events.
Artifacts are stored for 5 days, but only created for push events.
The commit also add rsync to the build dependencies of the Debian
package, because some of the tests need that.
The commit also disables the Travis Debian build instructions.