MAINT: Replace remainig Travis CI jobs with GitHub workflows
This commit adds a GitHub workflow to run nosetests, both with only the
basic requirements installed and with all the requirements installed.
The steps to setup the host system are in the script `setup-full.sh` in
the workflow scripts folder, so we don't have to clutter the workflow
definition.
The workflow also runs codecov.
Another workflow added in this commit builds the documentation using
sphinx.

The commit also adds a `skip_ci` check for tests, that skips the check
if it runs on a host that has the environment variable CI set to true.
This commit enables this check in the
intelmq/tests/bots/experts/national_cert_contact_certat/test_expert.py
because that one failes with GitHub workflows.
One check in intelmq/tests/lib/test_pipeline.py now skips not only on
Travis but on CIs in general.