DOC: clean up intelmq-api documentation
Remove the code blocks showing how to run intelmq-api wiht uwsgi and
gunicorn- thats not useful for endusers.
Rewrite the configuration paragraph and update the default values listed
for reference.
Mention that Debian uses another path for the session file.
Update the user configuration paragraph to mention
``intelmq-api-adduser`` instead of the hug command.