MAINT: Remove Malware Domain List feed
This commit removes the malwaredomainlist parser bot and all relevant
unit tests. It removes the relevant feed entry in the feeds.yaml file
and drops the malwaredomainlist bot entry in the BOTS file, in the
runtime.conf and in the pipeline.conf.
Last but not least it adds two methods to the upgrades.py removing bots
that use the malwaredomainlist parser module and pointing out collectors
that have the malwaredomain list url set. The upgrades are also tested
in the test_upgrade unit test.