BUG: cymru cap parser: adapt parsing of bruteforce
the format of category "bruteforce" changed
adapts the parsing

Previously, the bruteforce events only had the protocol in the comment, while most other categories had a (more or less) clear type-value syntax. Now, the bruteforce categories also uses the type-value syntax. That makes the parser a bit more bloated, as it needs to support both formats, the old and the new. Whereas I don't even know if this is a real switch, or just an additional format.

Interestingly the data no longer contains the application protocol (SSH in this case) :/

fixes certtools/intelmq#1794