ENH: Make intelmq compatible to yaml files
Luckily json is a subset of yaml, so we can simply use the yaml loader
to also load the legacy json files. For writing files, this commit
introduces an argument `yaml` to the write_configuration function that
enables storing files as yaml.
The commit also adds the yaml python package to the debian dependencies
and the python requirements.
The intelmq/__init__.py file now does a check to see if there is an old
runtime.conf and then moves that to runtime.yaml
The shipped intelmq/etc/runtime.conf, which was in json format, was
replaced by a intelmq/etc/runtime.yaml in YAML format.