BUG: lib/bot_debugger: fix setting the log level
Set bot's `logging_level` directly in `__init__` before the bot's initialization by changing the default value.

Rewrite `load_configuration_patch` by adapting it to the parameter and configuration rewrite.
If Runtime configuration is detected, the logging_level parameter is
 - inserted in all bot's parameters. bot_id is not accessible here, hence we add it everywhere
 - inserted in the global parameters (ex-defaults).
Maybe not everything is necessary, but we can make sure the logging_level is just everywhere where it might be relevant, also in the future.