TST: add a workflow to run ansible tests
This commit implements a github workflow that runs the ansible intelmq
tests on the checked out intelmq git repository. It pulls both the
intelmq repository and the repository with the ansible tests, then it
starts a Debian container and runs the ansible tests in the container.
For now the workflow comes with its own intelmq runtime file, which
should be integrated in the upstream ansible test repository (or even
be integrated in the default runtime.yaml shipped by intelmq) in the
future.