DOC: add license information to all the files
This commit adds license information to a lot of files and adds a
.reuse/dep5 file that lists the license information for some folders

The commit also changes the main license in setup.cfg from AGPL-3.0-only
to AGPL-3.0-or-later because only one file has the AGPL-3.0-only file as
license and multiple files have the AGPL-3.0-or-later in the license
header.

It also removes the cef_logo.png file, as there is no information about
the licese anywhere to be found. It is now included directly from the
website of the european union.

Closes #1633