ENH: Mixin & once behaviour
Mixins are initialized before the bot init, this is mandatory
for the mixins, because mixins have to be initialized before
the bot gets initialized. Otherwise the bot would raise errors
because mixins cant access data.

once() is a newly introduced function and will run once after bot
start. This function can be used for some after start cleanups or
pre-data collection from pipelines.

Signed-off-by: Sebastian Waldbauer <waldbauer@cert.at>