DOC: Document the licenses of all the files
This commit adds a license header or a license file to most of the
files, or documents the license in the .reuse/dep5 license file.

Some of the process was automated, first by listing all the files that
are not reuse lint compliant:
> reuse lint > ../reuse.lst
This list was then modified to remove metainformation and only list
filenames. Also a couple of filenames that need manual modification were
removed.

Then using git and reuse:
> for file in `cat ../reuse.lst`; do year=`git log --reverse --pretty="format:%ai" $file | head -1 | cut -d "-" -f 1`;  author=`git log --reverse --pretty="format:%an" $file|head -1`; reuse addheader --copyright="$author" --year="$year" --license="AGPL-3.0-or-later" --skip-unrecognised $file; done

Then the same process was repeated for files reuse does not recognize,
like csv and json files or REQUIREMENTS.txt files.