DOC: n6: add more illustrations
add two n6 images directly to the repository, as they are not displayed
on readthedocs otherwise: The other websites hosting the images block
loading images if the referer does not match a whitelist. we can't add a
noreferer HTML attribute in rst as well. the option left is to add the
files, that only implies adding the licensing information and the
AGPL-3.0 license text as well.

add two illustrations on the the flow n6 to intelmq and vice versa, own
work.

some textual improvements in the document itself.