Refactoring intelmqctl
This commit introduces a separate module for the processmanagers. The
module also contains a processmanager interface all the processmanager
should adhere to.

The commit also moves some of the commonly used values to the datatypes
module, so that therer is a central place for them to reside in.

Both the processmanagers and the intelmqctl command were cleaned up a
bit.