BUG: gethostbyname: handle numeric gaierrors_to_ignore
If the configuration for the gethostbyname expert contains

    gaierrors_to_ignore: -3

or if you enter a (single) numeric value in the manager, which automatically saves it as integer, not as string, the bot does not start with:

gethostbyname-1-expert: Bot initialization failed.
Traceback (most recent call last):
  File "/dev/intelmq/intelmq/lib/bot.py", line 209, in __init__
    self.init()
  File "/dev/intelmq/intelmq/bots/experts/gethostbyname/expert.py", line 43, in init
    ignore = ignore.split(',')
AttributeError: 'int' object has no attribute 'split'

A workaround is to encapsulate the value in quotes.

fixes certtools/intelmq#2072