FIX: Fixed re-reloading the harmonization file after each iteration
As of now each new message re-loads the actual harmonization conf,
which unfortuneatly results in a high CPU drain.

This commit adds the harmonization parameter to the function call,
so the harmonization is loaded from the internal python memory
allocated space & no more file read/write has to be done.

Fixes #2098

Signed-off-by: Sebastian Waldbauer <waldbauer@cert.at>