ENH: Add JSON parser function to Jinja2 templates
Jinja2 has a JSON serializer, but not a parser. This is the source of
many StackOverflow questions and bug reports to projects using Jinja2.
Add a function "from_json", which just calls "json.loads", to
templates so that they can handle the JSON string in the "output"
field of an event.