BUG: Fix line recovery and message dumping in ParserBot
ParserBot.process first iterates over all message in a parsed report and
catches the errors, saves the erroneous message and at the end dumps all
of them in one row.
The recover_line methods were not consistent in their use of the line
given as parameter and the use of the _current_line class member.
This lead to parsers always dumping the last message (_current_line
which was active at the time of the dumping) leading to data loss for
dumped messages for `recover_line_csv_dict` and
`recover_line_json_stream`.