FIX: handled bot's SyntaxError in intelmqclt check
The command 'intelmqctl check' can now handle SyntaxError
caused in bot's code without crashing. In addition,
the command got first unittests.

The 'utils.list_all_bots' command is adjusted to omit
modules with SyntaxErrors. This is consistent with the
'intelmqctl check' command behaviour, what is the only
one place this function is used.

closes: #2177