FIX: intelmqdump: respected logging_path settings
'intelmqdup' supports now configuration of logging_path,
both global and custom per-both. It can find and restore
dumps from all possibly configured locations. For readibility
reasons, the behaviour of showing just the filename in the
dumps list is left unchanged. If there is more than one
dump listed with the same filename, selecting them using
filename will throw an error, and the number should be used.

Unit testes for the command are introduced. For the purpose
of the change, just checking an empty dump file is enough.

closes: #1605