FIX: CSV parser bot preserves original line ending
CSV parser bot recognizes the original line ending
in the input report and tries to keep it in the
output. If input has multiple different line endings,
the \r\n is used. This is what Python csv module uses
by default.

Previously, line filtering was handled in a few places,
in case of CSV data it was before creating the handler object
and required rewriting the whole CSV data, with hardcoded line
endings. For proper handling line ending, this the filtering was
moved inside the handler and now it does not depend on any line
ending.

In addition, the filtering code was extracted to remove code
duplication and the CSV dict line parser was updated to use
the original line ending.