SMTPBatchOutput bot (#2253)
SMTPBatchOutput bot

Hi, we use this bot in CSIRT.cz since 2015 to inform our constituents about all recent incidents twice a week. Events are aggregated in a redis cache by the abuse e-mail and sent at once.

Since we are reaching the external organisations, we've put together some experience to tune up the format of the CSV so that people of various IT skills were able to open it in their spreadsheet software, we use GPG signing and pack the events into ZIP files as we sometimes hit the ceiling of an SMTP server.

We've recently upgraded the syntax to honour the IntelMQ 3, cleaned up the code, put some tips (like CRON setup) to bots.rst and be glad if the community would accept the bot.