ENH: Partitioning and rerun in schema generation
Those changes improves the capabilities of the database
schema creation. It's now easier to switch to the TimescaleDB
or use partitioning in PostgreSQL. Creating scripts can be
generated from the harmonisation, as previously just the most
basic version. If requested, creation commands get IF NOT EXISTS
or OR REPLACE directives to allow re-runing script.