Discover bots based on the entry points
The bot discovery algorithm was rewritten to use entry point
names instead of the file structure. This allow really easy
developing of packages with custom bots.

Not that the 'group' feature of entry points wasn't used
because we require bots to register callable scripts that
are later used to start bot's processes.