FIX: Correctly configure SSL context for AMQP
According to the Python documentation, in order to establish
a client-side connection the purpose of SERVER_AUTH is appropriate.
It instructs the SSL libraries to verify the authenticity
of the server. As of Python 3.10 [2], the internal configuration
used has been changed from the generic PROTOCOL_TLS to a specific
variant for client or server. Therefore, when using Python 3.10
or higher, the incorrect configuration of ssl.Purpose results
in an inability to establish a connection.

[1] https://docs.python.org/3/library/ssl.html#ssl.Purpose.SERVER_AUTH
[2] https://docs.python.org/3/library/ssl.html#ssl.PROTOCOL_TLS