Severity field in IDF
Severity is expected in IntelMQ for a long time and partially, it's
already used by e.g. ShadowServer reports. This implementation is
based on their understanding of the field, but with explicit
mentioning that operators could adjust it based on their knowledge.

This is not intended to be an ultimate severity classification,
but a help for first triage of recived events.

Close #2365